refactor(Weapon): prefab layout and script structure
