+++ CommunityProject/Assets/_Game/Prefabs/Weapons/Weapon_Base.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 219993bb6ae684db684a090b29c97d73
+guid: 58040a623dccf48739eb35e77740a26f
 PrefabImporter:
   externalObjects: {}
   userData:
