+++ CommunityProject/Assets/_Game/Scripts/Runtime/Weapons/BallisticWeapons.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 59874b16bdb74920b616b55d3f5a33cb
+timeCreated: 1674720858
