+++ CommunityProject/Assets/_Game/Scripts/Runtime/Weapons/BallisticWeapons/BallisticWeaponDiagnostics.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+
+namespace BoundfoxStudios.CommunityProject.Weapons.BallisticWeapons
+{
+	/// <summary>
+	/// Helper class for Gizmos, see Editor Assembly -> BallisticWeaponGizmos
+	/// </summary>
+	[AddComponentMenu(Constants.MenuNames.Weapons + "/" + nameof(BallisticWeaponDiagnostics))]
+	public class BallisticWeaponDiagnostics : MonoBehaviour
+	{
+		[field: SerializeField]
+		public BallisticWeapon Weapon { get; private set; }
+	}
+}
