+++ CommunityProject/Assets/_Game/Scripts/Runtime/Weapons/BallisticWeapons/ScriptableObjects.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 8db2578f2318407a8a09349b2f4ae592
+timeCreated: 1674720884
