--- CommunityProject/Assets/_Game/Scripts/Runtime/Weapons/Bullets.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 3d0e41d6c006480ea66dc3eb63cac01f
-timeCreated: 1674030572
